Add HTTP Exceptions

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { Repository } from 'typeorm';
@@ -9,7 +9,13 @@ import { UpdateUserDto } from './dto/update-user.dto';
 export class UsersService {
   constructor(@InjectRepository(User) private userRepository: Repository<User>) { }
 
-  createUser(user: CreateUserDto) {
+  async createUser(user: CreateUserDto) {
+    const findUser = await this.getUserByUsername(user.username);
+
+    if (findUser) {
+      return new HttpException('User already exists', HttpStatus.CONFLICT);
+    }
+
     const newUser = this.userRepository.create(user)
     return this.userRepository.save(newUser);
   }
@@ -18,19 +24,65 @@ export class UsersService {
     return this.userRepository.find();
   }
 
-  getUser(id: number) {
-    return this.userRepository.findOne({
+  async getUserById(id: number) {
+    const userFound = await this.userRepository.findOne({
       where: {
         id
       }
     })
+
+    if (!userFound) {
+      return new HttpException('User does not exist', HttpStatus.NOT_FOUND);
+    }
+
+    return userFound;
   }
 
-  deleteUser(id: number) {
-    return this.userRepository.delete({ id })
+  getUserByUsername(username: string) {
+    return this.userRepository.findOne({
+      where: {
+        username
+      }
+    })
   }
 
-  updateUser(id: number, user: UpdateUserDto) {
-    return this.userRepository.update({ id }, user)
+  async deleteUser(id: number) {
+    const userFound = await this.userRepository.findOne({
+      where: {
+        id
+      }
+    })
+
+    if (!userFound) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return this.userRepository.delete({ id });
+  }
+
+  async deleteUserPro(id: number) {
+    const result = await this.userRepository.delete({ id });
+
+    if (result.affected === 0) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return result;
+  }
+
+  async updateUser(id: number, user: UpdateUserDto) {
+    const userFound = await this.userRepository.findOne({
+      where: {
+        id
+      }
+    })
+
+    if (!userFound) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    const updateUser = Object.assign(userFound, user)
+
+    return this.userRepository.save(updateUser);
   }
 }
